+++ config/filament-trumbowyg.php
@@ -5,4 +5,5 @@ return [
     'tag_classes' => [],
     'change_active_dropdown_icon' => false,
     'semantic' => true,
+    'remove_format_pasted' => false,
 ];